breaking changes: New default AOT compatible dependency resolver, remove most AOT labels. (#1461)
* breaking changes: New AOT compatible dependency resolver, soften the AOT attributes

* Refactor tests to use target-typed new expressions and simplify method bodies

- Updated various test classes to utilize target-typed `new` expressions for cleaner instantiation.
- Simplified method bodies by removing unnecessary braces and return statements in async methods.
- Adjusted assertions to directly use the result of method calls where applicable.
- Ensured consistency across tests in handling null checks and service registrations.

* Add comprehensive unit tests for `FullLoggerExtensions` and related logging features

- Introduced new tests in `FullLoggerExtensionsTests` to validate behavior across all log levels (Debug, Info, Warn, Error, Fatal) for standard loggers, generic loggers, and exception-based loggers.
- Added coverage for `StaticLoggerTests` to verify constructor exception handling.
- Expanded `CoverageColorTests` to include validation tests for known color properties.

* Update `FullLoggerExtensionsTests` to standardize null function handling

- Replaced inline `null!` usage with `Func<string> nullFunc = null!` across all test cases.
- Ensured consistency when testing null argument exceptions for loggers at all levels (Debug, Info, Warn, Error, Fatal).

* refactor: improve dependency registration and cleanup within `NinjectDependencyResolver`
- Added null-check handling for contracts in registration methods to enhance robustness.
- Simplified and centralized logic for marking registrations and managing disposables.
- Updated comments and docstrings for clearer documentation and understanding.
- Refactored `GenericFirstDependencyResolver` based classes with improved thread-safety, state handling, and null/unwrapped results processing.

* refactor: improve contract handling and update resolver documentation

- Updated `HasRegistration` to handle null contracts correctly and align with Splat's expected behavior.
- Enhanced documentation in the README to clarify performance and isolation differences between resolvers.
- Added recommendations for resolver usage and detailed performance characteristics for better user guidance.

* Add back the features of the new container table

* Moved logging unit tests to its own assemblies, added missing tests unit tests.

Fixed other unit tests.

+++ src/Splat/Maths/PointMathExtensions.cs
@@ -57,10 +57,10 @@ public static class PointMathExtensions
     /// </summary>
     /// <param name="value">The point value to use for the calculation.</param>
     /// <returns>The normalized point.</returns>
-    public static PointF Normalize(this PointF value) => (float)value.Length() switch
+    public static PointF Normalize(this PointF value) => value.Length() switch
     {
         0.0f => value,
-        _ => new(value.X / (float)value.Length(), value.Y / (float)value.Length())
+        _ => new(value.X / value.Length(), value.Y / value.Length())
     };
 
     /// <summary>
